web/src/requirements.txt
Flask==0.12.2
Flask-Script==2.0.5
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.2
Flask-Migrate==2.0.4
Flask-Admin==1.5.0
alembic==0.9.3
